+++ core/src/main/scala/org/datatools/bigdatatypes/formats/SnakifyFormats.scala
@@ -1,10 +1,12 @@
 package org.datatools.bigdatatypes.formats
 
+import org.datatools.bigdatatypes.types.basic.SqlType
+
 /** Converts CamelCase field names to snake_case
   */
 trait SnakifyFormats extends Formats {
 
-  override def transformKey(name: String): String = name
+  override def transformKey[A <: SqlType](name: String, t: A): String = name
     .replaceAll("([A-Z]+)([A-Z][a-z])", "$1_$2")
     .replaceAll("([a-z\\d])([A-Z])", "$1_$2")
     .toLowerCase